WIP: Update UI to show issues by assignment
