+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+# Contributing
+
+More details required.....
+
+
+# Running the app from the source code
+
+```
+python ./pygitissue2todoist/PyGitIssue2Todoist.py  
+```
+
+# Testing
+
+To run the tests use
+
+```
+python -m unittest discover
+```
