+++ pygitissue2todoist/adapters/GitHubAdapter.py
@@ -36,6 +36,7 @@ class AbbreviatedGitIssue:
     issueTitle:   str = ''
     issueHTMLURL: str = ''
     body:         str = ''
+    prefix:       str = ''
     labels:       List[str] = field(default_factory=createLabelsFactory)
 
 
@@ -143,6 +144,13 @@ class GithubAdapter:
         simpleIssue.issueTitle   = fullGitIssue.title
         simpleIssue.issueHTMLURL = fullGitIssue.html_url
         simpleIssue.body         = fullGitIssue.body
+
+        # Used to display the organization and repository name, in the Assigned To Me issue list
+        try:
+            simpleIssue.prefix = f"[{fullGitIssue.repository.organization.login}/{fullGitIssue.repository.name}]"
+        except AttributeError:
+            simpleIssue.prefix = f"[{fullGitIssue.repository.name}]"
+
         for label in fullGitIssue.labels:
             gitLabel: Label = cast(Label, label)
             simpleIssue.labels.append(gitLabel.name)
@@ -157,9 +165,11 @@ class GithubAdapter:
         `gh search issues --assignee "@me" --include-prs`
 
         TODO: confirm how to include the PRs in the results
-        `gh search issues --assignee "@me" --include-prs`
+        `gh search issues --assignee "@me" --include-prs --state=open`
         """
-        openGitIssues: PaginatedList = self._github.search_issues(query='assignee:@me')
+        openGitIssues: PaginatedList = self._github.search_issues(
+            query='assignee:@me state:open',
+        )
         simpleGitIssues: AbbreviatedGitIssues = AbbreviatedGitIssues([])
 
         for openIssue in openGitIssues:
