+++ pygitissue2todoist/ui/TodoistPanel.py
@@ -34,9 +34,9 @@ from pygitissue2todoist.adapters.AbstractTodoistAdapter import AbstractTodoistAd
 from pygitissue2todoist.adapters.AdapterAuthenticationError import AdapterAuthenticationError
 from pygitissue2todoist.adapters.TodoistAdapter import CloneInformation
 from pygitissue2todoist.adapters.TodoistAdapter import GitIssueInfo
+# from pygitissue2todoist.adapters.TodoistAdapter import TodoistAdapter
+# from pygitissue2todoist.adapters.TodoistAdapterSingleProject import TodoistAdapterSingleProject
 from pygitissue2todoist.adapters.TodoistAdapter import TodoistAdapter
-from pygitissue2todoist.adapters.TodoistAdapterSingleProject import TodoistAdapterSingleProject
-from pygitissue2todoist.adapters.TodoistAdapterV2 import TodoistAdapterV2
 
 from pygitissue2todoist.general.PreferencesV2 import PreferencesV2
 
@@ -76,7 +76,7 @@ class TodoistPanel(BasePanel):
         #     self._todoistAdapter: AbstractTodoistAdapter = TodoistAdapterSingleProject(apiToken=self._apiToken)
         # else:
         #     self._todoistAdapter = TodoistAdapter(self._apiToken)
-        self._todoistAdapter = TodoistAdapterV2(self._apiToken)
+        self._todoistAdapter = TodoistAdapter(self._apiToken)
 
         self._taskList:         ListBox = cast(ListBox, None)
         self._createTaskButton: Button  = cast(Button, None)
@@ -192,6 +192,6 @@ class TodoistPanel(BasePanel):
                 # The following 2 already defined in init
                 self._apiToken       = PreferencesV2().todoistAPIToken
                 # self._todoistAdapter = TodoistAdapter(self._apiToken)
-                self._todoistAdapter = TodoistAdapterV2(self._apiToken)
+                self._todoistAdapter = TodoistAdapter(self._apiToken)
 
                 self._onCreateTaskClicked(event)    # Dang I hate recursion
